Endpoint fixes and now works with database
