+++ src/app/api/users/route.js
@@ -2,11 +2,10 @@ import { NextResponse } from "next/server";
 
 import db from "@/db/database";
 import bcrypt from "bcrypt";
-import { create } from "@mui/material/styles/createTransitions";
 
 // POST: /api/users/
 // Create a new user to the database given a username, uemail and upassword
-export async function POST(req) {
+export async function POST(request) {
 	const { username, uemail, upassword } = await request.json();
 
 	if (!username || !uemail || !upassword) {
@@ -26,6 +25,7 @@ export async function POST(req) {
 		);
 	}
 
+	// hardcoded salt rounds
 	const passwordHash = await bcrypt.hash(upassword, 12);
 
 	const createUser = db.prepare(
@@ -35,41 +35,6 @@ export async function POST(req) {
 
 	return NextResponse.json({
 		message: "User created successfully.",
-		userId: result.lastInsertRowid,
-	});
-}
-
-// GET: /api/users/<uid>
-// Retrieves all data for a given user's uid
-export async function GET(req, { params }) {
-	const { uid } = await params;
-
-	if (!uid) {
-		return NextResponse.json(
-			{ error: "Missing id field." },
-			{ status: 400 }
-		);
-	}
-
-	const user = db.prepare("SELECT * FROM users WHERE uid = ?").get(uid);
-	if (!user) {
-		return NextResponse.json(
-			{ error: "User does not exist." },
-			{ status: 404 }
-		);
-	}
-
-	// important to not send the password hash
-	return NextResponse.json({
-		message: "User found.",
-		user: {
-			uid: user.uid,
-			username: user.username,
-			ufirstname: user.ufirstname,
-			ulastname: user.ulastname,
-			uemail: user.uemail,
-			ucity: user.ucity,
-			urating: user.urating,
-		},
+		uid: result.lastInsertRowid,
 	});
 }
